requirements.txt
asgiref==3.5.2
autopep8==2.0.1
certifi==2022.12.7
charset-normalizer==2.1.1
Django==4.1.4
idna==3.4
pycodestyle==2.10.0
python-decouple==3.6
python-dotenv==0.21.0
requests==2.28.1
sqlparse==0.4.3
tomli==2.0.1
urllib3==1.26.13
